requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
gunicorn==21.2.0
flask-cors==4.0.0
python-dotenv==1.0.0
setuptools==69.0.3